CRQ-4 move test code for configSanityCheck to tests directory
